+++ packages/datasource-customizer/test/context/wrapper.test.ts
@@ -39,6 +39,7 @@ describe('RelaxedWrappers', () => {
     beforeEach(() => {
       caller = factories.caller.build();
       collection = factories.collection.build({
+        nativeDriver: 'my-native-driver',
         schema: {
           fields: {
             name: factories.columnSchema.build({ columnType: 'String' }),
@@ -46,6 +47,10 @@ describe('RelaxedWrappers', () => {
         },
       });
       relaxed = new RelatedCollection(collection, caller);
+    });
+
+    test('should forward nativeDriver calls', async () => {
+      expect(relaxed.nativeDriver).toBe('my-native-driver');
     });
 
     test('should call list when provided with a plain object', async () => {